Improved conversion between Unicode/ISO-5426.
Based on http://www.unicode.org/L2/L2000/00220-map-5426.pdf there were
some characters missing and some incorrect for the conversion to/from
Unicode to/from ISO-5426.

In ISO-5426, characters in the range 0xC0-0xDF are used as combining
characters to apply diacritics to the characters that _follow_ them. The
existing code handles these by outputting the normalized,
single-character versions when converting to unicode, and only
supporting normalized unicode input when converting from unicode.
This means that data input to UnicodeToIso5426 needs to be normalized to
get the correct result, so that was added.

Characters 0xC8 and 0xC9 (4/8, 4/9) map to the same unicode character,
u+0308 "combining diaeresis". When converting to ISO-5426, current code
always uses the 0xC8 vairent when converting the normalized version of
<some-char> + u+0308.  The reverse conversion back to unicode had no
support for 0xC9 characters even though this is legal. Now 0xC9
preceding a character will produce the same unicode character as 0xC8
preceding a character would.

Mappings for ISO-5426 characters 0xF2 and 0xF3 (7/2, 7/3) were missing
in both directions.

Mappings for ISO-5426 characters 0xA2, 0xAA, 0xB0, 0xB1, and 0xB2 (2/2,
2/10, 3/0, 3/1, 3/2) were either wrong or missing in one direction.